config

+++ app/(tabs)/profileTab.tsx
@@ -20,7 +20,7 @@ import { useFocusEffect } from "@react-navigation/native";
 import { onAuthStateChanged } from "firebase/auth";
 import { router } from 'expo-router'
 import { getAuth, updateEmail, updatePassword } from 'firebase/auth';
-
+import Config from 'react-native-config'; // Import Config to access env variables
 
 export default function TabFourScreen() {
   getAuth().onAuthStateChanged((user) => {
@@ -85,10 +85,12 @@ export default function TabFourScreen() {
   const [modalVisible, setModalVisible] = useState(false);
   const [newEmail, setNewEmail] = useState('');
   const [newPassword, setNewPassword] = useState('');
-  
+  console.log("API Base URL:", Config.API_BASE_URL);
+  const API_BASE_URL = Config.API_BASE_URL; // Get API URL from .env
+
   const handleUpdateCredentials = async () => {
     const user = auth.currentUser;
-    if (!user) return;
+    if (!user || !API_BASE_URL) return;
   
     try {
       // Validate Rutgers email
@@ -172,11 +174,12 @@ export default function TabFourScreen() {
   );
 
   const fetchProfile = async () => {
-    if (!user) return;
+    if (!user || !API_BASE_URL) return;
     setLoading(true);
     try {
+
       const token = await user.getIdToken();
-      const response = await fetch("http://127.0.0.1:5000/api/profile", {
+      const response = await fetch(`https://rumble-xe2g.onrender.com/api/profile`, {
         headers: { Authorization: token },
       });
       const data = await response.json();
@@ -191,11 +194,12 @@ export default function TabFourScreen() {
     setLoading(false);
   };
   const updateProfile = async () => {
-    if (!user) return;
+    if (!user || !API_BASE_URL) return;
     setLoading(true);
     try {
       const token = await user.getIdToken();
-      const response = await fetch("http://127.0.0.1:5000/api/update_profile", {
+      const response = await fetch(`https://rumble-xe2g.onrender.com/api/profile`, {
+
         method: "PUT",
         headers: {
           "Content-Type": "application/json",
@@ -241,7 +245,7 @@ export default function TabFourScreen() {
   const [pendingImageUpdate, setPendingImageUpdate] = useState(false);
 
   const uploadImage = async (uri: string) => {
-    if (!user) return;
+    if (!user || !API_BASE_URL) return;
     setLoading(true);
     try {
       const response = await fetch(uri);
@@ -286,11 +290,11 @@ useEffect(() => {
 }, [userProfile.profile.profilePictureUrl, pendingImageUpdate]);
 
   const deleteAccount = async () => {
-    if (!user) return;
+    if (!user || !API_BASE_URL) return;
     setLoading(true);
     try {
       const token = await user.getIdToken();
-      const response = await fetch("http://127.0.0.1:5000/api/delete_account", {
+      const response = await fetch(`${API_BASE_URL}/delete_account`, {
         method: "DELETE",
         headers: {
           Authorization: `${token}`,
